backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
langchain>=0.1.0
langgraph>=0.0.15
openai>=1.12.0
google-generativeai>=0.3.2
pydantic>=2.5.2
python-multipart>=0.0.6
typing-extensions>=4.9.0
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5 